+++ packages/client/hmi-client/src/workflow/ops/model-edit/tera-model-edit.vue
@@ -317,7 +317,6 @@ const initializeAceEditor = (editorInstance: any) => {
 // FIXME: Copy pasted in 3 locations, could be written cleaner and in a service
 const saveCodeToState = (code: string, hasCodeBeenRun: boolean) => {
 	const state = _.cloneDeep(props.node.state);
-	console.log(state);
 	state.hasCodeBeenRun = hasCodeBeenRun;
 
 	// for now only save the last code executed, may want to save all code executed in the future
